Arpeggio is working but bugs remain
As the title says arpeggios/note automatic manipulation is now working
but there are still some bugs. Namely the arpeggio needs to make a full
cycle otherwise it goes into the fault handler. This is probably due to
it accessing null memory in the linked list somewhere due to some
uninitialized section